refactor: payment summery before depositing
