+++ src/modules/Contract/components/PaymentSummary/index.module.scss
@@ -1,0 +1,50 @@
+@import 'src/styles/constants/_primitives.scss';
+
+.container {
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    gap: 1.5rem;
+    padding: 1.25rem 1rem 1.5rem 1rem;
+    border: 1px solid $color-grey-200;
+    border-radius: $radius-md;
+
+    .title {
+        font-size: 18px;
+        font-weight: 600;
+        line-height: 28px;
+        color: $color-grey-900;
+    }
+
+    .col {
+        display: flex;
+        flex-direction: column;
+        gap: 0.5rem;
+    }
+
+    .row {
+        display: flex;
+        align-items: center;
+        justify-content: space-between;
+        font-size: 14px;
+        font-weight: 400;
+        line-height: 20px;
+        color: $color-grey-700;
+
+        &--bold {
+            font-size: 16px;
+            font-weight: 500;
+            line-height: 24px;
+        }
+
+        &--bolder {
+            font-size: 24px;
+            font-weight: 600;
+            line-height: 32px;
+        }
+
+        &--success {
+            color: $color-success-600;
+        }
+    }
+}
