+++ src/modules/Contract/components/PaymentSummary/index.types.ts
@@ -1,0 +1,8 @@
+export interface PaymentSummaryProps {
+  amount: number;
+  sociousFee: number;
+  stripeFee: number;
+  total: number;
+  currency: string;
+  hasFeeDiscount: boolean;
+}
